initial refactor, move out of pkg
