+++ cmd/webapptool/subcommands/datastorage.go
@@ -6,7 +6,7 @@ import (
 	"io"
 	"net/http"
 
-	"github.com/dvo-dev/go-get-started/pkg/utils/requests"
+	"github.com/dvo-dev/go-get-started/utils/requests"
 	"github.com/spf13/cobra"
 )
