+++ customerrors/client.go
