+++ routes/datastorage.go
@@ -1,4 +1,4 @@
-package handlers
+package routes
 
 import (
 	"bytes"
@@ -8,9 +8,9 @@ import (
 	"net/http"
 	"strings"
 
-	"github.com/dvo-dev/go-get-started/pkg/customerrors"
-	"github.com/dvo-dev/go-get-started/pkg/datastorage"
-	"github.com/dvo-dev/go-get-started/pkg/responses"
+	"github.com/dvo-dev/go-get-started/customerrors"
+	"github.com/dvo-dev/go-get-started/responses"
+	"github.com/dvo-dev/go-get-started/services/datastorage"
 )
 
 // DataStorageHandler is a wrapper struct for `DataStorage` implementations,
